requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
filelock==3.6.0
huggingface-hub==0.4.0
idna==3.3
joblib==1.1.0
lmdb==1.3.0
numpy==1.22.3
packaging==21.3
Pillow==9.0.1
pyparsing==3.0.7
PyYAML==6.0
regex==2022.3.2
requests==2.27.1
sacremoses==0.0.47
six==1.16.0
tokenizers==0.11.6
torch==1.11.0
torchvision==0.12.0
tqdm==4.63.0
transformers==4.17.0
typing_extensions==4.1.1
urllib3==1.26.8
wincertstore==0.2

matplotlib~=3.5.1
scikit-learn~=1.0.2